+++ Api/Api/Repositories/Configuration/TimeseriesConfiguration.cs
@@ -12,7 +12,7 @@ namespace Api.Repositories.Configuration
             builder.HasKey(t => t.Id);
             builder.Property(t => t.Name).HasMaxLength(50).IsRequired();
             builder.Property(t => t.Timestamp).IsRequired();
-            builder.Property(t => t.Value).IsRequired().HasColumnType("decimal");
+            builder.Property(t => t.Value).IsRequired().HasColumnType("float");
         }
     }
 }
