+++ Api/Api/Repositories/Models/Timeseries.cs
@@ -7,6 +7,6 @@ namespace Api.Repositories.Models
         public Guid Id { get; set; }
         public string Name { get; set; }
         public DateTime Timestamp{ get; set; }
-        public decimal Value { get; set; }
+        public double Value { get; set; }
     }
 }
